+++ app/db.py
@@ -1,0 +1,1 @@
+database = []
